requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
openpyxl==3.1.2
anthropic==0.18.0
openai==1.12.0
google-cloud-vision==3.5.0
pydantic==2.5.3
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
